requirements.txt
openai-whisper
moviepy
imageio-ffmpeg
